requirements/base.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environ==1.0
Markdown==3.7
psycopg2==2.9.10
PyJWT==2.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1